flattened import structure
