+++ pyproject.toml
@@ -4,7 +4,7 @@ build-backend = "hatchling.build"
 
 [project]
 name = "seastate"
-version = "0.2.0beta00"
+version = "0.2.0b1"
 authors = [{ name = "Ivo Rivetta", email = "ivo@ivorivetta.com" }]
 description = "Collect ocean measurements for location and timeframe. Integrates NDBC and TidesAndCurrents datasources."
 readme = "README.md"
